rustdoc: Move inlining to its own module
